fix(enter): disable command on non-Linux platforms
